refactor: move all tile static methods out of tile geometry
